Some progress on driver and controller
Getting to the point to where we can run the program and test some. Did
some work on driver to link things together, added main frame to to
controller so it can add/remove components as needed (might need to be a
new class). Some work on selection view as well, implementing it as a
JPanel so it can be added to the main frame with other components.